bot/requirements.txt
python-telegram-bot[job-queue]>=20.0
SQLAlchemy[asyncio]>=2.0
asyncpg>=0.29.0
greenlet>=3.0.0
langchain>=0.1.10
langchain-google-genai>=2.0.0
langchain-community>=0.0.25
chromadb>=0.4.24
python-dotenv>=1.0.0
unstructured>=0.10.0
markdown
httpx
langchain-groq
